refactor(webfont-generator): extract generation pipeline
