+++ packages/webfont-generator/src/templates/html.rs
@@ -212,8 +212,9 @@ mod tests {
     use super::{build_html_context, build_html_registry, render_html_with_context};
     use crate::input::LoadedSvgFile;
     use crate::templates::SharedTemplateData;
+    use crate::types::ResolvedGenerateWebfontsOptions;
     use crate::util::{path_to_slashes, relative_path};
-    use crate::{FontType, GenerateWebfontsOptions, ResolvedGenerateWebfontsOptions};
+    use crate::{FontType, GenerateWebfontsOptions};
 
     use crate::test_helpers::{fixture_source_files, resolve_options, write_temp_template};
